Move controller back to instance.go
